Update patch set 2
Patch Set 2:

Improves things significantly, but doesn't completely fix them.

Going to look into playing with stretching the hsync porches and reducing the vsync porches to compensate.

These settings give a bitclk of 896.7672, which is slightly lower than the other panels, and one of the vsync porches is huge compared to most vidmode panels.

Patch-set: 2