Make the brcm ldisc driver compatible with libbt (use it for BT)
Kernel changes that allow to use the broadcom ldisc driver for combined
BT and FMRadio chips for BT using libbt. The major modification is the
interception of the hci configuration commands emitted by libbt and
their simulated answer to the upper layers. The configuration is done
already by the driver, while a reconfiguration would break the
functionality of the ldisc driver.
This modification requires that libbt is configured to communicate via
the ldisc's device node ("/dev/brcm_bt_drv"), while uim is wired to the
"real" uart port.

Required userspace modifications:
* "#define HCILP_INCLUDED false" in bdroid_buildcfg.h
* "UartPort = /dev/brcm_bt_dev" in bt_vendor.conf
* "UimUartPort = /dev/ttyHS0" in bt_vendor.conf (for msm_serial_hs)

Signed-off-by: Alexander Diewald <Diewi@diewald-net.com>
Change-Id: Ib1a551b47280d3881f71d12b0dcfd274ff7efc4b