remove msm8956-v1.1 1958 MHz node
The msm8956 runs with 1.8 GHz in the BIG cluster, if you want to run with more clock you should overvolt it. This clock is cause of random reboots in AOSP roms. Sol, remove it if you want to fix it. Or, if you want to overclock, add 50 ~ 75 mV in the regulator node of the BIG cpu corresponding to this clock.